Abstract class and text output
